requirements.txt
# Core dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.9
langchain-core>=0.1.7

# Graph processing
networkx>=3.0
faiss-cpu>=1.7.0  # Use faiss-gpu if GPU available

# Embeddings
sentence-transformers>=2.2.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep learning (optional, for local LLM)
torch>=1.9.0
transformers>=4.20.0

# Utilities
tiktoken>=0.5.0
jsonlines>=4.0.0
python-dotenv>=1.0.0

# Evaluation
anthropic>=0.18.0  # For Claude evaluation
google-generativeai>=0.3.0  # For Gemini evaluation

# Optional: Chinese LLM services
qianfan>=0.4.0  # Optional: For Baidu Qianfan LLM services

